Return type changed

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
     public double calculate(double heightMeter, double weightKg) {
-        return (weightKg / heightMeter / heightMeter);
+        return (int) (weightKg / heightMeter / heightMeter);
     }
 }
